refactor(fieldset): convert internal fieldset to typescript
