--- src/__internal__/fieldset/index.d.ts
@@ -1,1 +1,0 @@
-export { default } from "./fieldset";
