--- src/__internal__/fieldset/index.js
@@ -1,1 +1,0 @@
-export { default } from "./fieldset.component";
