+++ src/__internal__/fieldset/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./fieldset.component";
+export type { FieldsetProps } from "./fieldset.component";
